+++ build.sh
@@ -56,7 +56,7 @@ fi
 [ -f $HOME/.cargo/env ] && . $HOME/.cargo/env
 
 # add to path brew
-[ -f $HOME/.zprofile ] && . $HOME/.zprofile
+#[ -f $HOME/.zprofile ] && . $HOME/.zprofile
 
 build_blake2bf() {